+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+* Add `Grocer::PassbookNotification` for sending, well... Passbook
+  notifications. This kind of notification requires no payload.
 * Determining current environment is case-insensitive ([Oriol
   Gual](https://github.com/oriolgual))